feature-complete item/recipe DB; Perl 5.18
Upgraded Perl codebase to 5.18.  In this version, smartmatch and
given/when are flagged experimental due to planned overhauls in 5.20, so
I found every place I used them and converted to more basic code
structures.

This commit also sees a feature-complete item and recipe database
implementation.  I cut out the CHI file-caching because it's just not
thread-safe - but since I'm saving the raw JSON to the database anyway,
there's no need for me to keep a file cache anymore.