new API: skins! also MD5 fix
New API endpoints skins.json and skin_details.json were released to
support the wardrobe functionality.  A new attribute default_skin was
added to item_details.json.

Also fixed where the MD5 for items and recipes was always returning the
same value because it was working on the return value of
enchode::utf8(), which modifies its argument in place and only returns
1.  Now the data is encoded first, then the MD5 is calculated directly
on the encoded data.